// src/types/swagger/schema.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/notes/{noteId}": {
    put: {
      parameters: {
        path: {
          noteId: string;
        };
        body: {
          data: definitions["Note"];
        };
      };
      responses: {
        /** OK */
        200: {
          schema: definitions["Note"];
        };
        /** Bad request */
        400: unknown;
        /** Unauthorized */
        401: unknown;
        /** Not found */
        404: unknown;
      };
    };
  };
  "/notes": {
    get: {
      parameters: {
        query: {
          cursor?: string;
          limit?: number;
          labelId?: string;
        };
      };
      responses: {
        /** OK */
        200: {
          schema: definitions["GetNotesResponse"];
        };
        /** Bad request */
        400: unknown;
        /** Unauthorized */
        401: unknown;
        /** Not found */
        404: unknown;
      };
    };
    post: {
      parameters: {
        body: {
          data: definitions["Note"];
        };
      };
      responses: {
        /** OK */
        200: {
          schema: definitions["Note"];
        };
        /** Bad request */
        400: unknown;
        /** Unauthorized */
        401: unknown;
        /** Not found */
        404: unknown;
      };
    };
  };
  "/labels": {
    get: {
      responses: {
        /** OK */
        200: {
          schema: definitions["GetLabelsResponse"];
        };
        /** Bad request */
        400: unknown;
        /** Unauthorized */
        401: unknown;
        /** Not found */
        404: unknown;
      };
    };
    post: {
      parameters: {
        body: {
          data: definitions["Label"];
        };
      };
      responses: {
        /** OK */
        200: {
          schema: definitions["Label"];
        };
        /** Bad request */
        400: unknown;
        /** Unauthorized */
        401: unknown;
        /** Not found */
        404: unknown;
      };
    };
  };
  "/labels/{labelId}": {
    put: {
      parameters: {
        path: {
          labelId: string;
        };
        body: {
          data: definitions["Label"];
        };
      };
      responses: {
        /** OK */
        200: {
          schema: definitions["Label"];
        };
        /** Bad request */
        400: unknown;
        /** Unauthorized */
        401: unknown;
        /** Not found */
        404: unknown;
      };
    };
  };
}

export interface definitions {
  Note: {
    _id?: string;
    title?: string;
    content?: string;
    labels?: string[];
    user?: string;
    archived?: boolean;
    binned?: boolean;
  };
  Label: {
    _id?: string;
    name?: string;
  };
  GetLabelsResponse: {
    data?: definitions["Label"][];
  };
  GetNotesResponse: {
    data?: definitions["Note"][];
    cursor?: string;
    hasMore?: boolean;
  };
}

export interface operations {}

export interface external {}